Move private headers into the internal directory
